backend/requirements.txt
# FastAPI Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Supabase - Using latest stable version
supabase==2.28.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
pydantic[email]>=2.11.7  # Changed from ==2.10.6 to >=2.11.7
email-validator==2.3.0

# AI Integration
groq==0.12.0

# File Upload
cloudinary==1.41.0

# Real-time Communication
python-socketio==5.12.0

# Environment Variables
python-dotenv==1.0.1

# Utilities
requests==2.32.3
aiofiles==24.1.0